Feature

+++ app/Http/Controllers/Admin/CategoriesController.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
+use App\Models\Post;
 use App\Models\Subdomain;
-use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 
 class CategoriesController extends Controller
@@ -76,8 +77,6 @@ class CategoriesController extends Controller
     {
         $subdomains = Subdomain::all()->pluck('name', 'id');
         $current_domain = $category->subdomain->id;
-//       dd($default);
-
 
         return view('admin.moderation.categories.edit')
             ->with('category', $category)
@@ -107,10 +106,20 @@ class CategoriesController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy(Category $category)
     {
+        $post = Post::all()->pluck('category_id', 'id')->toArray();
+
+        if (in_array($category->id, $post)) {
+            return redirect()->back()->with('error', 'Помилка! Category привязанa до Post');
+        }
+
         $category->delete();
+
+        return redirect()->route('category.index')
+            ->with('success', "Category {$category->name} success remove");
     }
 }
+
